Package name change: changing to kainos package name tp prevent cluster conflicts
Package name change: removing integration files

Reverting dependency exclusion on 1.7.7

Package name change: adding correct test exclusions
