+++ src/main/scala/com/kainos/spark/avro/AvroOutputWriterFactory.scala
@@ -14,18 +14,19 @@
  * limitations under the License.
  */
 
-package com.databricks.spark.avro
+package com.kainos.spark.avro
 
-import org.apache.spark.sql.SQLContext
-import org.apache.spark.sql.sources.{HadoopFsRelation, HadoopFsRelationProvider}
+import org.apache.hadoop.mapreduce.TaskAttemptContext
+import org.apache.spark.sql.sources.{OutputWriter, OutputWriterFactory}
 import org.apache.spark.sql.types.StructType
 
-private[avro] class DefaultSource extends HadoopFsRelationProvider {
 
-  def createRelation(sqlContext: SQLContext,
-      paths: Array[String],
-      dataSchema: Option[StructType],
-      partitionColumns: Option[StructType],
-      parameters: Map[String, String]): HadoopFsRelation =
-    new AvroRelation(paths, dataSchema, partitionColumns, parameters)(sqlContext)
+private[avro] class AvroOutputWriterFactory(schema: StructType,
+    recordName: String,
+    recordNamespace: String) extends OutputWriterFactory {
+
+  override def newInstance(path: String,
+      dataSchema: StructType,
+      context: TaskAttemptContext): OutputWriter =
+    new AvroOutputWriter(path, context, schema, recordName, recordNamespace)
 }
