+++ src/test/scala/com/kainos/spark/avro/AvroLogicalGenerator.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.databricks.spark.avro
+package com.kainos.spark.avro
 
 import java.io.File
 import java.math.BigDecimal
