+++ src/test/scala/com/kainos/spark/avro/AvroSuite.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.databricks.spark.avro
+package com.kainos.spark.avro
 
 import java.io.{File, FileNotFoundException}
 import java.nio.ByteBuffer
@@ -281,7 +281,7 @@ class AvroSuite extends FunSuite with BeforeAndAfterAll {
     sqlContext.sql(
       s"""
          |CREATE TEMPORARY TABLE avroTable
-         |USING com.databricks.spark.avro
+         |USING com.kainos.spark.avro
          |OPTIONS (path "$episodesFile")
       """.stripMargin.replaceAll("\n", " "))
 
@@ -303,7 +303,7 @@ class AvroSuite extends FunSuite with BeforeAndAfterAll {
     // get the same values back.
     TestUtils.withTempDir { tempDir =>
       val name = "AvroTest"
-      val namespace = "com.databricks.spark.avro"
+      val namespace = "com.kainos.spark.avro"
       val parameters = Map("recordName" -> name, "recordNamespace" -> namespace)
 
       val avroDir = tempDir + "/namedAvro"
@@ -404,14 +404,14 @@ class AvroSuite extends FunSuite with BeforeAndAfterAll {
       sqlContext.sql(
         s"""
            |CREATE TEMPORARY TABLE episodes
-           |USING com.databricks.spark.avro
+           |USING com.kainos.spark.avro
            |OPTIONS (path "$episodesFile")
         """.stripMargin.replaceAll("\n", " "))
       sqlContext.sql(
         s"""
            |CREATE TEMPORARY TABLE episodesEmpty
            |(name string, air_date string, doctor int)
-           |USING com.databricks.spark.avro
+           |USING com.kainos.spark.avro
            |OPTIONS (path "$tempEmptyDir")
         """.stripMargin.replaceAll("\n", " "))
